+++ cicl_zadanie_6.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int decnumber(unsigned long x) {
-    int i = 0;
+    size_t i = 0;
     if (x == 0) return 1;
     while (x > 0) {
         i++;
@@ -30,7 +30,7 @@ int main() {
     scanf("%lu", &y);
     if(x > y) printf("error\n");
     else {
-        for (unsigned long i = x; i <= y; i++) {
+        for (size_t i = x; i <= y; i++) {
             if(decnumber(i) == 6) {
                 if(first_3_sum(i) == second_3_sum(i))
                     printf("%lu ", i);
